delete old saved claims job (#1852)
* delete file in before destroy callback for persistent attachment

* delete old claims job

* add job to scheduler

* add delete claims indexes

* make spec have success state for new submission

* lint

* typo

* allow skipping validations in specs

* put saved claim job in namespace

* lint
