+++ config/sidekiq_scheduler.yml
@@ -19,6 +19,11 @@ EducationForm::CreateSpoolSubmissionsReport:
   cron: "15 4 * * MON-FRI America/New_York"
   description: "Send the daily report to the call center about spool file submissions"
 
+CentralMail::DeleteOldClaims:
+  cron: "30 2 * * * America/New_York"
+  class: CentralMail::DeleteOldClaims
+  description: "Clear out central mail claims older than 2 months"
+
 DeleteOldApplications:
   cron: "0 0 * * * America/New_York"
   class: EducationForm::DeleteOldApplications
