+++ db/migrate/20180416231107_add_delete_claims_indexes_to_saved_claims.rb
@@ -1,0 +1,8 @@
+class AddDeleteClaimsIndexesToSavedClaims < ActiveRecord::Migration
+  disable_ddl_transaction!
+
+  def change
+    add_index(:saved_claims, [:created_at, :type], algorithm: :concurrently)
+    add_index(:central_mail_submissions, :state, algorithm: :concurrently)
+  end
+end
